# wsgi-scripts/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: オムライス API
  description: This is a Talent Management System API server.
  contact:
    name: オムライス
    email: miyahara.44112@XXXXX.co.jp
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://172.22.128.114
paths:
  /employees/search:
    post:
      tags:
        - employee
      summary: Search employee' results
      description: Get employee's search results
      operationId: get_employees_search_results
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/body"
            examples:
              example_ALLSearchCondition:
                value:
                  employeeId: s44112
                  employeeName: 宮原祥太
                  dept:
                    - BI2_旅行BI部
                  joiningYearFrom: {}
                  joiningYearTo: {}
                  title:
                    - "9"
                  grade:
                    - B2
                  recruitmentClass:
                    - "1"
                  leaveFlg:
                    - "1"
                  gender:
                    - "1"
                  birthdateFrom: {}
                  birthdateTo: {}
              example_PartSearchCondition:
                value:
                  employeeId: s44112
                  employeeName: 宮原祥太
                  title:
                    - "9"
                  grade:
                    - B2
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employees"
              examples:
                example-1:
                  value:
                    - employeeId: s44112
                      employeeName: 宮原祥太
                      joiningYear: {}
                      title:
                        - "9"
                      grade:
                        - B2
                      recruitmentClass:
                        - "1"
                      gender:
                        - "1"
                      birthdate: {}
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
      x-openapi-router-controller: swagger_server.controllers.employee_controller
  /employees/{employeeId}:
    get:
      tags:
        - employee
      summary: Get an employee info
      description: Get an employee info
      operationId: get_employees_employee_id
      parameters:
        - name: employeeId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            maxLength: 7
            minLength: 1
            type: string
            example: s44112
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeInfo"
              examples:
                example-1:
                  value:
                    employees:
                      - employeeId: s44112
                        employeeName: 宮原祥太
                        joiningYear: {}
                        title: "9"
                        grade: B2
                        recruitmentClass: "1"
                        gender: "1"
                        birthdate: {}
                    employeesAddress:
                      company: 株式会社XXXXX
                      mailAddress: test@XXX.com
                    attendTrainings:
                      - trainingName: データハンドリング研修
                        startDate: {}
                        satisfactionScore: 5
                    certificationInfo:
                      - certificationId: c000001
                        certificationName: 基本情報技術者試験
                        summary: データハンドリングの基礎を学ぶ研修
                        getDate: {}
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
      x-openapi-router-controller: swagger_server.controllers.employee_controller
...
  /employees:
    get:
      tags:
        - employee
      summary: Get All employees
      description: Get Employees Lists
      operationId: get_employees
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employees"
              examples:
                example-1:
                  value:
                    - employeeId: s44112
                      employeeName: 宮原祥太
                      joiningYear: {}
                      title: "9"
                      grade: B2
                      recruitmentClass: "1"
                      gender: "1"
                      birthdate: {}
        "400":
          description: Bad Request
        "404":
          description: Not Found
      x-openapi-router-controller: swagger_server.controllers.employee_controller
components:
  schemas:
    Employees:
      uniqueItems: true
      type: array
      items:
        $ref: "#/components/schemas/Employees_inner"
      x-examples: {}
...
    EmployeeInfo:
      type: object
      properties:
        employees:
          $ref: "#/components/schemas/Employees"
        employeesAddress:
          $ref: "#/components/schemas/EmployeeInfo_employeesAddress"
        attendTrainings:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeInfo_attendTrainings"
        certificationInfo:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeInfo_certificationInfo"
      example:
        certificationInfo:
          - summary: summary
            certificationId: certificationId
            certificationName: certificationName
            getDate: {}
          - summary: summary
            certificationId: certificationId
            certificationName: certificationName
            getDate: {}
        employeesAddress:
          mailAddress: mailAddress
          company: company
        employees:
          - joiningYear: joiningYear
            employeeName: employeeName
            birthdate: {}
            gender: gender
            grade: grade
            recruitmentClass: recruitmentClass
            employeeId: employeeId
            title: title
          - joiningYear: joiningYear
            employeeName: employeeName
            birthdate: {}
            gender: gender
            grade: grade
            recruitmentClass: recruitmentClass
            employeeId: employeeId
            title: title
        attendTrainings:
          - trainingName: trainingName
            satisfactionScore: 0.8008281904610115
            startDate: {}
          - trainingName: trainingName
            satisfactionScore: 0.8008281904610115
            startDate: {}
...
    body:
      type: object
      properties:
        employeeId:
          type: string
        employeeName:
          type: string
        dept:
          type: array
          items:
            type: string
        joiningYearFrom:
          type: string
          format: date
        joiningYearTo:
          type: string
          format: date
        title:
          type: array
          items:
            type: string
        grade:
          type: array
          items:
            type: string
        recruitmentClass:
          type: array
          items:
            type: string
        leaveFlg:
          type: array
          items:
            type: string
        gender:
          type: array
          items:
            type: string
        birthdateFrom:
          type: string
          format: date
        birthdateTo:
          type: string
          format: date
...
    Employees_inner:
      type: object
      properties:
        employeeId:
          type: string
        employeeName:
          type: string
        dept:
          type: string
        joiningYear:
          type: string
          format: date
        title:
          type: string
        grade:
          type: string
        recruitmentClass:
          type: string
        gender:
          type: string
        birthdate:
          type: string
          format: date
      example:
        joiningYear: joiningYear
        employeeName: employeeName
        birthdate: {}
        gender: gender
        grade: grade
        recruitmentClass: recruitmentClass
        employeeId: employeeId
        title: title
...
    EmployeeInfo_employeesAddress:
      type: object
      properties:
        company:
          type: string
        mailAddress:
          type: string
      example:
        mailAddress: mailAddress
        company: company
    EmployeeInfo_attendTrainings:
      type: object
      properties:
        trainingName:
          type: string
        startDate:
          type: string
          format: date
        attendance_state:
          type: string
        satisfactionScore:
          type: number
      example:
        trainingName: trainingName
        attendance_state: "1"
        satisfactionScore: 0.8008281904610115
        startDate: {}
    EmployeeInfo_certificationInfo:
      type: object
      properties:
        certificationId:
          type: string
        certificationName:
          type: string
        summary:
          type: string
        getDate:
          type: string
          format: date
      example:
        summary: summary
        certificationId: certificationId
        certificationName: certificationName
        getDate: {}
...
  securitySchemes: {}